Add settings upgrade step.
